Get rid of the local_run concept
