+++ codalab/worker/docker_image_manager.py
